build: don't use $PKG_CONFIG for wayland-scanner
I made a mistake in the previous commit — $PKG_CONFIG should not be
used to find wayland-scanner, because when cross-compiling that'll
look for the cross-compiled wayland-scanner.  I missed this because I
tested by compiling from glibc to musl on the same architecture, so I
was testing a special case where my system could actually execute the
cross-compiled wayland-scanner.

Fixes: f6a8203 ("build: fix cross-compilation with pkg-config")